+++ README.md
@@ -8,25 +8,28 @@ review-fix-1
 сервис для шеринга вещей
 
 ### Выполненные задачи
-**1. Создание каркаса приложения**
+**1. Дальнейшая проработка каркаса приложения**
 
-**2. Реализация модели данных (с спользованием описанного шаблона в доп. материалах к ТЗ)**
+**2. Реализация модели данных (с использованием описанного шаблона в доп. материалах к ТЗ)**
 * сущность "пользователь" (User)
 * сущность "вещь" (Item)
 * сущность "бронирование" (Booking)
 * сущность "запрос" (Request)
-* сущность "отзыв" (Review)
+* сущность комментарий (Comment)
 
-**3. Разработка части веб-слоя**
+**3. Разработка веб-слоя**
 * для сущностей пользователя (UserController)
 * для сущности "Вещь" (ItemController)
+* для комментариев к вещи после заказа
+* для создания заказов на бронирование вещей
 
-**4. Выделение DTO-классов и мапперов, соответствующих модели, для реализации веб-слоя**
+**4. Дальнейшая проработка DTO-классов и мапперов, соответствующих модели, для реализации веб-слоя**
 
-**5. Реализация хранения данных в оперативной памяти (in-memory, структура данных - HashMap)**
-* в качестве ключа целочисленный идентификатор
+**5. Реализация хранения данных в БД**
+* использование JPA
+* удалена реализация хранилища в памяти
 
-**5. Тестирование с помощью данной Postman-коллекции**
+**5. Тестирование с помощью данной Postman-коллекции СП-14**
 
-## Исправления и уточнения по ревью и обсуждению
-- вторая итерация + оптимизированы Responce Entity при обработки исключений
+## Исправления и уточнения по ревью и обсуждению СП14
+- третья итерация
