--- src/main/java/ru/practicum/shareit/config/AppConfig.java
@@ -1,7 +1,0 @@
-package ru.practicum.shareit.config;
-
-import org.springframework.context.annotation.Configuration;
-
-@Configuration
-public class AppConfig {
-}
