+++ src/main/java/ru/practicum/shareit/request/ItemRequestDtoMapper.java
@@ -10,7 +10,7 @@ import ru.practicum.shareit.request.model.ItemRequest;
 /**
  * ItemRequest <=> ItemRequestDto Mapstruct-маппер <p>
  * ТЗ-13 <p>
- * @implNote expect Sprint add-item-requests for more details
+ * @implNote expect Sprint 15 add-item-requests for more details
  */
 @Mapper(nullValuePropertyMappingStrategy = NullValuePropertyMappingStrategy.IGNORE,
         componentModel = "spring")
