+++ src/main/java/ru/practicum/shareit/user/UserDtoMapper.java
@@ -1,9 +1,6 @@
 package ru.practicum.shareit.user;
 
-import org.mapstruct.InheritConfiguration;
-import org.mapstruct.Mapper;
-import org.mapstruct.MappingTarget;
-import org.mapstruct.NullValuePropertyMappingStrategy;
+import org.mapstruct.*;
 import ru.practicum.shareit.user.dto.UserDto;
 import ru.practicum.shareit.user.model.User;
 
@@ -20,5 +17,11 @@ public interface UserDtoMapper {
     User fromDto(UserDto userDto);
 
     @InheritConfiguration
-    User update(UserDto source, @MappingTarget User destination);
+    @Mapping(target = "user.email", expression = "java(notNullBlankSource(dto.getEmail(), user.getEmail()))")
+    @Mapping(target = "user.name", expression = "java(notNullBlankSource(dto.getName(), user.getName()))")
+    void update(UserDto dto, @MappingTarget User user);
+
+    default String notNullBlankSource(String source, String target) {
+        return source != null && !source.isBlank() ? source : target;
+    }
 }
