+++ src/main/java/ru/practicum/shareit/user/UserService.java
@@ -1,79 +1,49 @@
 package ru.practicum.shareit.user;
 
-import lombok.RequiredArgsConstructor;
-import lombok.extern.slf4j.Slf4j;
-import org.springframework.stereotype.Service;
-import org.springframework.transaction.annotation.Transactional;
-import ru.practicum.shareit.exception.*;
 import ru.practicum.shareit.user.dto.UserDto;
-import ru.practicum.shareit.user.model.User;
 
 import java.util.List;
-import static java.util.stream.Collectors.toList;
-import static java.lang.String.format;
-
-import static ru.practicum.shareit.util.Constants.SUCCESS_DELETE_MESSAGE;
 
 /**
- * сервис-слой для обработки данных пользователей <p>
- * ТЗ-13
+ * интерфейс сервис-слоя для обработки данных вещах для шаринга <p>
+ * ТЗ-13 <p>
+ * CRUD-функционал
  */
-@Slf4j
-@Service
-@RequiredArgsConstructor
-public class UserService implements UserServing {
+public interface UserService {
 
-    private final UserDtoMapper userMapper;
+    /**
+     * добавление пользователя
+     * @param dto экз. DTO класса для пользователя
+     * @return экз. DTO для добавленного  пользователя
+     */
+    UserDto addUser(UserDto dto);
 
-    private final UserRepository userStorage;
+    /**
+     * частичное изменение данных пользователя
+     * @param userId идентификатор сохраненного пользователя
+     * @param dto DTO для пользователя <p>
+     * частично заполненные поля
+     * @return экз. DTO для добавленного  пользователя
+     */
+    UserDto patch(Long userId, UserDto dto);
 
-    @Transactional
-    @Override
-    public UserDto addUser(UserDto dto) {
-        User user = userMapper.fromDto(dto);
-        User created = userStorage.save(user);
-        return userMapper.toDto(created);
-    }
+    /**
+     * получение DTO для пользователя из хранилища
+     * @param userId идентификатор сохраненного пользователя
+     * @return экз. DTO для пользователя из хранилища
+     */
+    UserDto getById(Long userId);
 
-    @Transactional
-    @Override
-    public UserDto patch(Long userId, UserDto dto) {
-        User user = userStorage.findById(userId).orElseThrow(
-                () -> {
-                    log.info("Service error reading User with id {}", userId);
-                    throw new StorageErrorException(format("Service error reading User with id %d", userId));
-                }
-        );
-        userMapper.update(dto, user);
-        return userMapper.toDto(user);
-    }
+    /**
+     * получение списка DTO для всех пользователей из хранилища
+     * @return список DTO
+     */
+    List<UserDto> getAll();
 
-    @Override
-    public UserDto getById(Long userId) {
-        User user = userStorage.findById(userId).orElseThrow(
-                () -> {
-                    log.info("User#id {} not found", userId);
-                    throw new NotFoundException(format("User#id %d not found", userId));
-                }
-        );
-        return userMapper.toDto(user);
-    }
-
-    @Override
-    public List<UserDto> getAll() {
-        return userStorage.findAll().stream()
-                          .map(userMapper::toDto)
-                          .collect(toList());
-    }
-
-    @Transactional
-    @Override
-    public String deleteById(Long userId) {
-        if (userStorage.existsById(userId)) {
-            userStorage.deleteById(userId);
-            log.info("deleted item with id {}", userId);
-            return SUCCESS_DELETE_MESSAGE;
-        }
-        throw new BadRequestException(format("Error deleting User#id %d", userId));
-    }
+    /**
+     * удаление пользователя из хранилища
+     * @param userId идентификатор сохраненного пользователя
+     * @return DTO для пользователя (уудаленный пользователь)
+     */
+    String deleteById(Long userId);
 }
